backend/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fastapi==0.115.8
gunicorn==23.0.0
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.1.0
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.44.0
whitenoise==6.8.2
xgboost==2.1.4
